+++ src/lib.rs
@@ -82,7 +82,7 @@
 //!         image_size: Size::new(8, 8),
 //!         image_data_len: 192,
 //!         channel_masks: None,
-//!         row_order: RowOrder::BottomUp
+//!         row_order: RowOrder::BottomUp,
 //!     }
 //! );
 //!
@@ -110,6 +110,7 @@
 #![deny(missing_docs)]
 #![deny(missing_debug_implementations)]
 
+mod color_table;
 mod dynamic_bmp;
 mod header;
 mod pixels;
@@ -155,6 +156,10 @@ where
                 // 4 bytes per pixel.
                 // This check could be improved by using the bit masks available in BMP headers
                 // with version >= 4, but we don't currently parse this information.
+            } else if (raw_bmp.color_bpp() == Bpp::Bits1 || raw_bmp.color_bpp() == Bpp::Bits8)
+                && raw_bmp.color_table().is_some()
+            {
+                // Allow 1BPP and 8BPP images with color tables to be mapped to other color types.
             } else {
                 return Err(ParseError::MismatchedBpp(raw_bmp.color_bpp().bits()));
             }